+++ lib/filesaverplus_linux_stub.dart
@@ -1,0 +1,3 @@
+class FileSaverPlusLinux {
+  static void registerWith() {}
+}