+++ lib/filesaverplus_web_stub.dart
@@ -1,0 +1,3 @@
+class FileSaverPlusWeb {
+  static void registerWith(dynamic registrar) {}
+}