+++ lib/filesaverplus_windows_stub.dart
@@ -1,0 +1,3 @@
+class FileSaverPlusWindows {
+  static void registerWith() {}
+}